feat(reorg): reorganized files, renamed form component to basic info form component
