+++ client/src/pages/dashboard/dashboard-content/training/training.js
