+++ client/src/pages/dashboard/dashboard-content/training/training.test.js
